Add a testing harness for PKC. Move native pkc tests to a separate file.
Use testing harness in pkc tests.

Add failure tests for `contract_logic`.

Add more native pkc tests.

post rebase build fixes.

Fix new tests.

Use hash idx 0 for call stack item hashing.

Make clang tidy happy.

Fix circuit js tests.

clang tidy ://

Comment out stuff that depends on pedersen update in noir.

revert hash change.

Disable test.

forgot to diable.

Remove redundant call stack check in pkc inner. (was already commented out)

rebase fixes.

Test fix.

fix test.

test fix :/

Remove explicit srs path in pkc.

fix path.

Check no errors for pkc success test cases.
